moved load_board fn for ease of tests
